# ENH,BENCH: Optimize floor_divide for VSX4/Power10