This PR:
- Optimizes floor_divide (arr // arr) for VSX4
- Adds a benchmark routine for floor_divide (arr // arr)
- Adds an enhancement note as requested in my previous PR (#20821)

Some comments:
- I used raw VSX intrinsics since the universal intrinsics only supports vector division by a scalar
- I did not include ulonglong since this datatype does not benefit from this contribution
- The compiler was not able to vectorize the unsigned types (arr // arr) code automatically

See below some benchmarking information:

```
       before           after         ratio
     [6077afd6]       [4dc613ef]
     <main>           <p10_enh_intdiv>
-      60.1±0.1μs       57.2±0.1μs     0.95  ArrayDivInt(int64, 10000)
-     18.2±0.04μs       17.3±0.1μs     0.95  ArrayDivInt(uint8, 10000)
-     5.91±0.01ms      5.58±0.01ms     0.94  ArrayDivInt(int32, 1000000)
-        1.63±0ms      1.52±0.01ms     0.94  ArrayDivInt(uint8, 1000000)
-        6.96±0ms      6.48±0.01ms     0.93  ArrayDivInt(int8, 1000000)
-      60.6±0.4μs       56.3±0.5μs     0.93  ArrayDivInt(int32, 10000)
-      69.5±0.3μs       64.0±0.2μs     0.92  ArrayDivInt(int8, 10000)
-     24.3±0.02μs      21.2±0.03μs     0.87  ArrayDivInt(uint32, 10000)
-        2.25±0ms         1.94±0ms     0.86  ArrayDivInt(uint32, 1000000)
-      61.6±0.2μs       36.3±0.5μs     0.59  ScalarDivInt(int64, -43)
-      61.6±0.2μs       36.2±0.5μs     0.59  ScalarDivInt(int64, 43)
-      60.9±0.1μs       22.4±0.4μs     0.37  ScalarDivInt(int64, -8)
-      60.9±0.2μs       22.2±0.2μs     0.36  ScalarDivInt(int64, 8)
```